Added in Reform Local Government decision
This decision checks for whether non-existent nations ruled by the owner of said nations' capitals have the same political reforms. If not, it will change every political reform of the nonexistent nation(s) to match the ruler's political reforms.
This is to fix the issue of released nations changing gov type after just a few months due to mismatched political reforms as it is illogical for a dictatorship to release a democracy and vice versa.